requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
prefect>=3.0.0,<4.0.0

# ── Data ────────────────────────────────────────────────────────────────────
pandas==2.0.3
numpy==1.24.4
pyarrow==14.0.2

# ── GCP ─────────────────────────────────────────────────────────────────────
gcsfs==2024.10.0
google-cloud-storage>=2.14.0,<3.0.0
google-cloud-secret-manager>=2.18.0,<3.0.0

# ── AWS ────────────────────────────────────────────────���────────────────────
boto3>=1.34.0,<2.0.0
aiosignal==1.3.1
aiohttp==3.8.5

# ── DB (kept for compatibility) ─────────────────────ฟ───────────────────────
sqlalchemy==1.4.52
psycopg2-binary==2.9.9
pymongo==4.3.3
dnspython==2.6.1

# ── Misc ────────────────────────────────────────────────────────────────────
gitpython>=3.1.40,<4.0.0
opensearch-py==2.3.0
requests-aws4auth==1.1.1
jsonpath_ng==1.5.3